@@ -6,8 +6,12 @@
   "main": "index.html",
   "scripts": {
     "prod": "NODE_ENV=production gulp",
-    "dev": "gulp",
+    "prod:nopack": "NODE_ENV=production gulp --nopack",
+    "dev": "gulp --dev",
+    "dev:nopack": "gulp --dev --nopack",
     "standalone": "NODE_ENV=standalone gulp",
+    "standalone:nopack": "NODE_ENV=standalone gulp --nopack",
+    "nopack": "gulp --nopack",
     "install": "NODE_ENV=production gulp",
     "postinstall": "find node_modules/ -name '*.info' -type f -delete"
   },
@@ -21,9 +25,12 @@
     "npm": ">=2.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "babel-loader": "^8.0.0",
     "breakpoint-sass": "^2.7.0",
     "browser-sync": "^2.17.5",
-    "css-loader": "^0.25.0",
+    "css-loader": "^1.0.0",
     "del": "^2.2.2",
     "fabricator-assemble": "^1.2.0",
     "fancy-log": "^1.3.2",
@@ -33,25 +40,29 @@
     "gmaps-marker-clusterer": "^1.2.2",
     "gulp": "^4.0.0",
     "gulp-autoprefixer": "^6.0.0",
+    "gulp-babel": "^8.0.0",
     "gulp-csso": "^3.0.0",
     "gulp-if": "^2.0.0",
     "gulp-imagemin": "^4.0.0",
+    "gulp-minify": "^3.1.0",
     "gulp-sass": "^4.0.0",
     "gulp-sourcemaps": "^2.6.4",
     "gulp-svg-sprite": "^1.4.0",
     "html-webpack-plugin": "^2.24.1",
+    "jquery": "^1.12.4",
     "jquery-match-height": "^0.7.0",
     "jquery-shorten-plus": "git://github.com/rayestrada/jquery.shorten.git",
     "jquery.scrollto": "^2.1.2",
     "jquery.viewport": "^0.2.0",
     "lazysizes": "^4.0.0",
     "load-google-maps-api-2": "^1.0.2",
+    "merge-stream": "^1.0.1",
     "owl.carousel": "^2.2.0",
     "path": "^0.12.7",
     "standardize-sass": "^1.0.0",
     "sticky-kit": "git://github.com/leafo/sticky-kit.git",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.23.0",
     "url-loader": "^0.5.7",
-    "webpack": "^3.0.0"
+    "webpack": "^4.18.0"
   }
 }
