@@ -4,17 +4,25 @@
  *
  */
 
-(function ($) {
+($ => {
   // DOC READY
-  $(function () {
+  $(() => {
 
     // Assign target selectors.
-    var $nav = $('#navigation > ul.menu');
-    var $trig = $('#menu-trigger > a');
-    var $toMainContentLink = $('#top');
-    var $backToTopLink = $('#bottom');
-    var $mobileMenuClass = 'js-show-menu';
-    var $kids = $nav.children();
+    const $nav = $('#navigation > ul.menu');
+    const $trig = $('#menu-trigger > a');
+    const $toMainContentLink = $('#top');
+    const $backToTopLink = $('#bottom');
+    const $mobileMenuClass = 'js-show-menu';
+    const $kids = $nav.children();
+
+    // Removes focus class and focus from all menu items.
+    const outofmenu = () => {
+      $kids.each(function () {
+        $(this).removeClass('js-focus');
+        $nav.find('a').blur();
+      });
+    }
 
     // Add class for top level menu items.
     $kids.addClass('js-top-level-menu-item');
@@ -26,13 +34,13 @@
     });
 
     // Tabbing out of the end menu items.
-    $nav.find('a.js-last-menu-item').keydown(function (e) {
+    $nav.find('a.js-last-menu-item').keydown(e => {
       // Tabbed out no shift.
       if (e.which == 9 && !e.shiftKey) {
         outofmenu();
       }
     });
-    $nav.find('a.js-first-menu-item').keydown(function (e) {
+    $nav.find('a.js-first-menu-item').keydown(e => {
       // Shift tabbed out.
       if (e.which == 9 && e.shiftKey) {
         outofmenu();
@@ -41,9 +49,7 @@
 
     // Remove focus from menu when you are clicking outside of it.
     $(document).on('click', outofmenu());
-    $nav.on('click', function (e) {
-      e.stopPropagation();
-    });
+    $nav.on('click', e => e.stopPropagation());
 
     // Add class on focus.
     $nav.find('a').on({
@@ -51,32 +57,24 @@
         $(this).closest('.js-top-level-menu-item').addClass('js-focus')
           .siblings().removeClass('js-focus');
       },
-      mouseenter: function () {
+      mouseenter: () => {
         // Let normal hovers override.
         outofmenu();
       }
     });
 
-    // Removes focus class and focus from all menu items.
-    function outofmenu() {
-      $kids.each(function () {
-        $(this).removeClass('js-focus');
-        $nav.find('a').blur();
-      });
-    }
-
     /**
      * Skip links
      */
     $toMainContentLink.on({
-      focus: function () {
+      focus: () => {
         $trig.removeClass('js-return-check');
       },
-      click: function () {
+      click: () => {
         $trig.addClass('js-return-check');
       }
     });
-    $backToTopLink.on('focus', function () {
+    $backToTopLink.on('focus', () => {
       // Scroll back to top link into view.
       window.scrollTo(0, document.body.scrollHeight);
     });
