@@ -10,19 +10,19 @@
  *
  */
 
-module.exports = function throttle(fn, threshhold, scope) {
+const throttle = (fn, threshhold, scope) => {
   threshhold || (threshhold = 100);
-  var last;
-  var deferTimer;
+  let last;
+  let deferTimer;
   return function () {
-    var context = scope || this;
+    const context = scope || this;
 
-    var now = +new Date;
-    var args = arguments;
+    const now = +new Date;
+    const args = arguments;
     if (last && now < last + threshhold) {
       // hold on to it
       clearTimeout(deferTimer);
-      deferTimer = setTimeout(function () {
+      deferTimer = setTimeout(() => {
         last = now;
         fn.apply(context, args);
       }, threshhold);
@@ -33,3 +33,5 @@ module.exports = function throttle(fn, threshhold, scope) {
     }
   };
 }
+
+export default throttle;
